requirements_few_shot.txt
absl-py==2.2.1
addict==2.4.0
aiohappyeyeballs==2.4.4
aiohttp==3.10.11
aiosignal==1.3.1
alabaster==0.7.13
astor==0.8.1
asttokens==3.0.0
async-timeout==5.0.1
attrs==25.3.0
autograd==1.7.0
autograd-gamma==0.5.0
babel==2.17.0
backcall==0.2.0
cachetools==5.5.2
certifi==2022.12.7
charset-normalizer==2.1.1
click==8.1.8
colorama==0.4.6
comet_ml==3.49.7
comm==0.2.2
configobj==5.0.9
contourpy==1.1.1
cycler==0.12.1
debugpy==1.8.14
decorator==5.2.1
diffusers==0.23.1
docutils==0.20.1
duckdb==1.2.1
dulwich==0.22.4
ecos==2.0.14
einops==0.8.1
einops-exts==0.0.4
et_xmlfile==2.0.0
everett==3.1.0
executing==2.2.0
faiss-cpu==1.8.0.post1
filelock==3.13.1
fonttools==4.57.0
formulaic==1.1.1
frozenlist==1.5.0
fsspec==2024.6.1
ftfy==6.2.3
google-auth==2.38.0
google-auth-oauthlib==1.0.0
graphlib_backport==1.1.0
grpcio==1.70.0
h5py==3.11.0
huggingface-hub==0.29.3
idc-index==0.8.6
idc-index-data==20.0.3
idna==3.4
imageio==2.35.1
imagesize==1.4.1
importlib_metadata==8.5.0
importlib_resources==6.4.5
interface-meta==1.3.0
ipykernel==6.29.5
ipython==8.12.3
jedi==0.19.2
Jinja2==3.1.4
joblib==1.4.2
jsonschema==4.23.0
jsonschema-specifications==2023.12.1
jupyter_client==8.6.3
jupyter_core==5.7.2
kiwisolver==1.4.7
lazy_loader==0.4
lifelines==0.27.8
lightning==2.3.3
lightning-utilities==0.11.9
loguru==0.7.3
Markdown==3.7
markdown-it-py==3.0.0
MarkupSafe==2.1.5
matplotlib==3.7.5
matplotlib-inline==0.1.7
mdurl==0.1.2
mpmath==1.3.0
multidict==6.1.0
munch==4.0.0
nest-asyncio==1.6.0
networkx==3.0
nmslib==2.1.1
numexpr==2.8.6
numpy==1.24.1
nvidia-cublas-cu12==12.1.3.1
nvidia-cuda-cupti-cu12==12.1.105
nvidia-cuda-nvrtc-cu12==12.1.105
nvidia-cuda-runtime-cu12==12.1.105
nvidia-cudnn-cu12==8.9.2.26
nvidia-cufft-cu12==11.0.2.54
nvidia-curand-cu12==10.3.2.106
nvidia-cusolver-cu12==11.4.5.107
nvidia-cusparse-cu12==12.1.0.106
nvidia-nccl-cu12==2.18.1
nvidia-nvjitlink-cu12==12.8.93
nvidia-nvtx-cu12==12.1.105
nystrom-attention==0.0.14
oauthlib==3.2.2
opencv-python==4.11.0.86
openpyxl==3.1.5
osqp==1.0.3
packaging==24.2
pandas==2.0.3
parso==0.8.4
pexpect==4.9.0
pickleshare==0.7.5
pillow==10.2.0
pkgutil_resolve_name==1.3.10
platformdirs==4.3.6
pretrainedmodels==0.7.4
prompt_toolkit==3.0.51
propcache==0.2.0
protobuf==5.29.4
psutil==7.0.0
ptyprocess==0.7.0
pure_eval==0.2.3
pyarrow==17.0.0
pyasn1==0.6.1
pyasn1_modules==0.4.2
pybind11==2.6.1
Pygments==2.19.1
pyparsing==3.1.4
python-box==6.1.0
python-dateutil==2.9.0.post0
pytorch-lightning==2.3.0
pytorch_toolbelt==0.8.0
pytz==2025.2
PyWavelets==1.4.1
PyYAML==6.0.2
pyzmq==26.4.0
referencing==0.35.1
regex==2024.11.6
requests==2.28.1
requests-oauthlib==2.0.0
requests-toolbelt==1.0.0
rich==14.0.0
rpds-py==0.20.1
rsa==4.9
s5cmd==0.2.0
safetensors==0.5.3
scikit-image==0.21.0
scikit-learn==1.3.2
scikit-survival==0.22.2
scipy==1.10.1
seaborn==0.13.2
semantic-version==2.10.0
sentry-sdk==2.26.1
setproctitle==1.3.5
simplejson==3.20.1
six==1.17.0
snowballstemmer==2.2.0
Sphinx==7.1.2
sphinx-click==6.0.0
sphinxcontrib-applehelp==1.0.4
sphinxcontrib-devhelp==1.0.2
sphinxcontrib-htmlhelp==2.0.1
sphinxcontrib-jsmath==1.0.1
sphinxcontrib-qthelp==1.0.3
sphinxcontrib-serializinghtml==1.1.5
stack-data==0.6.3
sympy==1.13.1
tensorboard==2.14.0
tensorboard-data-server==0.7.2
tensorboardX==2.6.2.2
threadpoolctl==3.5.0
tifffile==2023.7.10
timm==1.0.3
tokenizers==0.20.3
torch==2.1.2
torch-geometric==2.6.1
#torch_scatter==2.1.2
#torch_sparse==0.6.18
torchaudio==2.1.2
torchmetrics==1.5.2
torchvision==0.16.2
tornado==6.4.2
tqdm==4.67.1
traitlets==5.14.3
transformers==4.46.3
triton==2.1.0
typing_extensions==4.12.2
tzdata==2025.2
urllib3==1.26.13
wcwidth==0.2.13
Werkzeug==3.0.6
wrapt==1.17.2
wurlitzer==3.1.1
yarl==1.15.2
zipp==3.20.2
